NetCache API: implemented generation of mirrored blob keys when using mirrored NetCache servers; introduced NetCache keys in CompoundID format. JIRA: CXX-4474, TRY-1781
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@60576 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/sample/app/netcache/netcache_client_sample.cpp
@@ -399,7 +399,8 @@ void CSampleNetCacheClient::DemoIWriterIReader(void)
             // available, without an error.  Therefore, bytes_read must
             // always be used after the read instead of read_size.
             size_t bytes_read;
-            ERW_Result rw_res = reader->Read(my_buf, read_size, &bytes_read);
+            ERW_Result rw_res = reader->Read(my_buf,
+                    (size_t) read_size, &bytes_read);
             my_buf[bytes_read] = '\0';
 
             if (rw_res == eRW_Success) {
